test: Refactor placeholder tests to fail with error
