+++ features/resume/internal-api/src/androidTest/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internalapi/PlaceholderInstrumentationTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.testwrapper
+package uk.gov.onelogin.criorchestrator.features.resume.internalapi
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Assert.fail
