--- features/resume/internal/src/androidTest/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/EmptyInstrumentationTest.kt
@@ -1,12 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.resume.internal
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class EmptyInstrumentationTest {
-    @Test
-    fun emptyTest() {
-    }
-}
