+++ features/resume/internal/src/androidTest/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/PlaceholderInstrumentationTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.testwrapper
+package uk.gov.onelogin.criorchestrator.features.resume.internal
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Assert.fail
