--- sdk/internal/src/androidTest/kotlin/uk/gov/onelogin/criorchestrator/sdk/internal/ExampleInstrumentedTest.kt
@@ -1,17 +1,0 @@
-package uk.gov.onelogin.criorchestrator.sdk.internal
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Assert.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun stubInstrumentationTest() {
-        assertEquals(
-            2,
-            1 + 1,
-        )
-    }
-}
