+++ sdk/public-api/src/androidTest/kotlin/uk/gov/onelogin/criorchestrator/sdk/publicapi/PlaceholderInstrumentationTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.testwrapper
+package uk.gov.onelogin.criorchestrator.sdk.publicapi
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Assert.fail
